Add tests for resolvers
